fixed bug that sometimes added new genes that were clones of existing ones

+++ src/main/java/org/encog/neural/neat/NEATPopulation.java
@@ -670,7 +670,7 @@ public class NEATPopulation extends AbstractNEATPopulation {
 				.flatMap(g -> ((NEATGenome)g).getNeuronsChromosome().stream())
 				.count();
 
-		return (numLinks + numNodes) / genomes.size();
+		return (numLinks + numNodes) / (double)genomes.size();
 	}
 
 }
